weather_agent/pyproject.toml
[project]
name = "ai-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
